feat: process device data from mqtt server
